Paper on Comparision of Semi-Empirical Model and XGBoost Model for CYGNSS Soil Moisture Estimation
SMAP and SMOS space missions are reliable sources of soil moisture, providing coarse resolution data (a temporal resolution of more than three days and spatial resolutions of 36 km and 25 km, respectively). The resolution of soil moisture can be improved by using one of the successful GNSS-R missions, CYGNSS (Cyclone GNSS). CYGNSS, with eight microsatellites in its constellation, captures a large volume of observations in the equatorial regions. Each microsatellite can receive four reflections per second with varying spatial resolutions from 0.5 km x 3.5 km to 25 km x 25 km and a temporal resolution of 1 day. In this paper, a soil moisture estimation model was made for CYGNSS satellite data so that the CYGNSS observations could be used differently to improve the soil moisture resolution. Since vegetation has always remained a problem for estimating soil moisture, this paper aims to account for vegetation attenuation to improve the model's performance. The SMAP soil moisture was highly correlated with the CYGNSS surface reflectivity (correlation of 81.6% on barren land and correlation of 81.2% on vegetated land). Ulaby's Water Cloud Model worked well with the CYGNSS data (1.5 GHz) by ignoring the vegetation attenuation because it depends on frequency. The model was more biased towards the higher frequency signals (more than 8 GHz). By adding the leaf area index as the fractional contribution of backscatter from vegetation to Ulaby's Water Cloud Model, a lot of changes were made that helped solve this problem. The modified water cloud model has performed well compared to Ulaby's model in the vegetated and built-in regions, with more than 75% correlations and RMSE less than 0.01 m3/m3 on average. Furthermore, the XGBoost regression algorithm was used, yielding correlations of up to 87% between predicted and SMAP soil moisture with a RMSE of 0.001 m3/m3 and a relative error of less than 0.2%.